Add samples and README
