+++ .dotnet/src/Custom/Files/OpenAIFileInfo.cs
@@ -7,7 +7,7 @@ public partial class OpenAIFileInfo
     public string Id { get; }
     public OpenAIFilePurpose Purpose { get; }
     public string Filename { get; }
-    public long Size { get; }
+    public long? Size { get; }
     public DateTimeOffset CreatedAt { get; }
 
     internal OpenAIFileInfo(Internal.Models.OpenAIFile internalFile)
