+++ .dotnet/tests/Samples/Embeddings/Sample01_SimpleEmbedding.cs
@@ -1,0 +1,31 @@
+﻿using NUnit.Framework;
+using OpenAI.Embeddings;
+using System;
+
+namespace OpenAI.Samples
+{
+    public partial class EmbeddingSamples
+    {
+        [Test]
+        [Ignore("Compilation validation only")]
+        public void Sample01_SimpleEmbedding()
+        {
+            EmbeddingClient client = new("text-embedding-3-small", Environment.GetEnvironmentVariable("OpenAIClient_KEY"));
+
+            string description =
+                "Best hotel in town if you like luxury hotels. They have an amazing infinity pool, a spa,"
+                + " and a really helpful concierge. The location is perfect -- right downtown, close to all "
+                + " the tourist attractions. We highly recommend this hotel.";
+
+            Embedding embedding = client.GenerateEmbedding(description);
+            ReadOnlyMemory<float> vector = embedding.Vector;
+
+            Console.WriteLine($"Dimension: {vector.Length}");
+            Console.WriteLine($"Floats: ");
+            for (int i = 0; i < vector.Length; i++)
+            {
+                Console.WriteLine($"  [{i}] = {vector.Span[i]}");
+            }
+        }
+    }
+}
